Move BluePrint libraries close to their associated structs.
